Burningsvideo 0.52 Linux Warnings
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@6158 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CSoftwareTexture2.cpp
@@ -350,7 +350,7 @@ void CSoftwareTexture2::calcDerivative()
 	}
 
 	//preCalc mipmap texel center boundaries
-	for (s32 i = 0; i < array_size(MipMap); ++i)
+	for (size_t i = 0; i < array_size(MipMap); ++i)
 	{
 		CSoftwareTexture2_Bound& b = TexBound[i];
 		if (MipMap[i])
